### 2025-07-15 #8589 @danielskatz
👋 @hacknus - thanks for your submission. However, your paper is far too long. JOSS expects papers of 250-1000 words, and yours is almost 3000. If you feel all the content in your paper is needed, you will need to find another journal to publish in.

If you want to publish in JOSS, you need to cut your paper. Perhaps there is content that you can put in the repo and just point to from the paper?  You can use the command `@editorialbot check repository` to check the paper length (and some other things), and the command `@editorialbot generate pdf` after making changes to the .md file to make a new PDF. editorialbot commands need to be the first entry in a new comment.
