+++ paper/paper.md
@@ -31,13 +31,11 @@ bibliography: paper.bib
 
 THz time-domain spectroscopy (TDS) is a fast-growing field with applications to perform non-destructive studies of
 material properties [@neu_tutorial_2018].
-Different sources of THz radiations have been implemented in commercial products, e.g. photo-conductive antennas. The
-pulses can either be measured after passing through (transmission spectrum) or after being reflected (reflection
+The pulses can either be measured after passing through (transmission spectrum) or after being reflected (reflection
 spectrum) by a sample and are recorded in the time domain. By transforming the
 acquired trace into frequency domain (Fourier space), the magnitude and phase can be extracted to
 investigate the complex refractive index and absorption coefficient of the sample.
-By placing either the sample or the optical setup on a moving stage the sample can be imaged in 2D. Analysing these
-images pixel by pixel or by selecting a region of interest (ROI) without an interactive user interface can be tedious.
+By placing either the sample or the optical setup on a moving stage the sample can be imaged in 2D.
 
 ![THz Image Explorer icon.\label{fig:icon}](icon.png){#id .class width=20%}
 
@@ -45,8 +43,7 @@ We developed an interactive graphical user interface (GUI), written
 in [Rust](https://www.rust-lang.org) [@matsakis2014rust], to aid
 investigating acquired 2D scans. The
 application implements the dotTHz standard [@lee_dotthz_2023] and is platform independent and open-source, thus making
-it
-easier to maintain and increasing its reach.
+it easier to maintain and increasing its reach.
 
 ![THz Image Explorer screenshot.\label{fig:screenshot}](screenshot.png)
 
@@ -75,132 +72,25 @@ The GUI uses [egui](https://www.egui.rs), an immediate-mode GUI library for rust
 and [bevy](https://bevy.org) [@bevyengine], a game
 engine used for rendering.
 
-The GUI thread handles all the user input and displaying of plots and other window elements. The configuration values
-set in the GUI are sent to the Data thread
+The configuration values set in the GUI are sent to the data thread
 via multiple-producer-multi-consumer (MPMC) channels.
-The Data thread then handles the computation of the applied filters.
-The output of the computation is then shared via mutexes with the GUI thread.
-The entire thread communication is handled with the `ThreadCommunication` struct which holds many `Arc<RwLock<T>>` or
-`crossbeam_channel::Sender<T>`/`crossbeam_channel::Receiver<T>`.
+The data thread then handles the computation of the applied filters.
+The outputs of the computation are then shared via mutexes with the GUI thread.
+The entire thread communication is handled with the `ThreadCommunication` struct.
 
 The structure of the software architecture is shown in figure \ref{fig:software-architecture}.
 
 ![Software Architecture.\label{fig:software-architecture}](thz-image-explorer.drawio.png){width=80% .center}
 
 For each filter, an entry in the `filter_data_pipeline` vector is created, which contains the dataset. Each filter is
-assigned an input and output index. This is memory intensive, but for the size of usual THz TDS datasets (tens to
-hundreds of MB) it is acceptable. This structure defines the processing pipeline in a clean and easily extendable way.
-
-# Installation
-
-## Pre-built Bundles
-
-Pre-built bundles are available for each release on [GitHub](https://github.com/unibe-icelab/thz-image-explorer) for
-
-- macOS (`.app` bundle for x86 and Apple Silicon)
-- Linux (executable and `.deb` for x86)
-- Windows (`.exe` and `.msi` for x86)
-
-These bundles should work out of the box, but on macOS you might need to remove the quarantine flag after downloading by
-running the following command:
-
-```shell
[…]
