+++ README.MD
@@ -17,8 +17,8 @@ A sample scan of a THz time-domain 2D imaging system can be found in the `sample
 
 - [Features](#features)
 - [Installation](#installation)
-  - [Download Pre-built Executables](#download-pre-built-executables)
-  - [Compile from Source](#compile-from-source)
+    - [Download Pre-built Executables](#download-pre-built-executables)
+    - [Compile from Source](#compile-from-source)
 - [Usage](#usage)
 - [Custom Filters](#custom-filters)
 - [PSF File Generation](#psf-file-generation)
@@ -41,11 +41,13 @@ A sample scan of a THz time-domain 2D imaging system can be found in the `sample
 
 ### Download Pre-built Executables
 
-[Binary bundles](https://github.com/unibe-icelab/thz-image-explorer/releases) are available for Linux, macOS, and Windows.
+[Binary bundles](https://github.com/unibe-icelab/thz-image-explorer/releases) are available for Linux, macOS, and
+Windows.
 
 #### macOS
 
 If you see the message `"THz Image Explorer is damaged and cannot be opened."`, run:
+
 ```sh
 xattr -rd com.apple.quarantine THz\ Image\ Explorer.app
 ```
@@ -53,6 +55,7 @@ xattr -rd com.apple.quarantine THz\ Image\ Explorer.app
 #### Linux
 
 Install dependencies:
+
 ```sh
 sudo apt-get install libxcb-render0-dev libxcb-shape0-dev libxcb-xfixes0-dev libxkbcommon-dev libssl-dev libasound2-dev
 ```
@@ -67,13 +70,14 @@ sudo apt-get install libxcb-render0-dev libxcb-shape0-dev libxcb-xfixes0-dev lib
    sudo apt-get install libxcb-render0-dev libxcb-shape0-dev libxcb-xfixes0-dev libxkbcommon-dev libssl-dev libasound2-dev
    ```
 3. Install bundling tools:
-   - Linux/macOS: `cargo install cargo-bundle`
-   - Windows: `cargo install cargo-wix`
+    - Linux/macOS: `cargo install cargo-bundle`
+    - Windows: `cargo install cargo-wix`
 4. Build:
-   - Linux/macOS: `cargo bundle --release`
-   - Windows: `cargo wix -p thz-image-explorer`
+    - Linux/macOS: `cargo bundle --release`
+    - Windows: `cargo wix -p thz-image-explorer`
 
-> **Note:** [cargo-bundle](https://github.com/burtonageo/cargo-bundle) only supports Linux and macOS. For Windows, use [cargo-wix](https://github.com/volks73/cargo-wix).
+> **Note:** [cargo-bundle](https://github.com/burtonageo/cargo-bundle) only supports Linux and macOS. For Windows,
+> use [cargo-wix](https://github.com/volks73/cargo-wix).
 
 #### macOS Plugin
 
@@ -87,30 +91,107 @@ Requires HDF5 to be installed system-wide.
 
 Example data is available in the `sample_data` directory of this repository.
 
-- For files larger than 100 x 100 x 2000 datapoints, use **downscaling** for better responsiveness.
-- Edit `.thz` file meta-data in the Meta Data Editor:
+The window is structured with the time domain trace and frequency domain spectrum for the selected pixel (default is
+0,0) in the center panel. A different tab showing the optical properties (refractive index and absorption coefficient)
+can be selected, as
+well as a tab containing an interactive 3D viewer.
+The left side-panel contains the intensity plot of the 2D scan along with the meta-data editor. The right
+side-panel contains the possible filters with configuration settings.
+A pixel can be (de-)selected by clicking inside the intensity plot.
[…]
