### 2025-07-15 72ddc7f
add some info about dotTHz standard
