### 2025-08-04 #8589 @fruzsinaagocs
:wave: @hacknus, thank you for your submission!

👋 @DanNixon & @lazear, would any of you be willing to review this submission for JOSS? Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.
