### 2025-08-05 401230b
fix polygon averaging (x/y)
