### 2025-08-05 6117d5e
show position in mm of selected pixel
