### 2025-08-05 #8589 @fruzsinaagocs
👋 @jsdodge, would you (or someone from your research group) be willing to review this submission for JOSS? Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.
