### 2025-08-05 #8589 @jsdodge
Sure, I can do it. I'm assuming that most of the work has already been done?
