### 2025-08-05 #8589 @fruzsinaagocs
I'll close the pre-review issue and start the review now.
