### 2025-08-05 #8754 @editorialbot
[REVIEW]: THz Image Explorer - An Interactive Cross-Platform Open-Source THz Image Analysis Tool
**Submitting author:** @hacknus (<a href="http://orcid.org/0000-0002-7916-2592">Linus Leo Stöckli</a>)
**Repository:** https://github.com/unibe-icelab/thz-image-explorer
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.1
