### 2025-08-06 #8754 @jsdodge
Hi @fruzsinaagocs , I ran into an issue with compiling the application from the source code. Am I correct in understanding that I should work with the authors to address this problem as part of the review? I didn't see anything about this in the documentation for reviewers, so sorry if I missed it.
