### 2025-08-06 #8589 @jsdodge
I realized that I confused this software project with a completely different one that I had just finished reviewing for pyOpenSci, which is why I suggested that most of the work had already been done in the pyOpenSci review. I'm still happy to review—this is right up my alley—but I thought I should explain my earlier comment.
