### 2025-08-07 82d857f
fix compilation instructions
