### 2025-08-09 0726779
fix color mapping (red-hot, blue-cold) and bw
