### 2025-08-09 6d94440
add git commit to info / settings window
