### 2025-08-12 #8754 @jsdodge
Hi @hacknus , yes, I was missing CMake. I am able to build it from source now.
