### 2025-08-12 95e419d
add readme badges
