+++ Cargo.toml
@@ -73,6 +73,9 @@ uuid = "1.17.0"
 downcast-rs = "2.0.1"
 vtkio = "0.7.0-rc1"
 
+[dev-dependencies]
+approx = "0.5.1"
+
 [target.'cfg(not(target_os = "macos"))'.dependencies]
 egui-file-dialog = { git = "https://github.com/hacknus/egui-file-dialog", branch = "sort_by_metadata", features = ["information_view"] }
 

[…]
