### 2025-08-14 7b47ccd
add community guidelines
