+++ Cargo.toml
@@ -61,15 +61,15 @@ reqwest = { version = "0.12", default-features = false, features = ["blocking", 
 semver = { version = "1.0.24", optional = true }
 filter_macros = { path = "filter_macros" }
 cancellable_loops = { path = "cancellable_loops" }
-ctor = "0.4"
+ctor = "0.5"
 once_cell = "1.21"
-rayon = "1.10.0"
+rayon = "1.11"
 rustfft = "6.2.0"
 num-complex = "0.4"
 interp1d = "0.2.0"
 chrono = "0.4"
 num-traits = "0.2.19"
-uuid = "1.17.0"
+uuid = "1.18.0"
 downcast-rs = "2.0.1"
 vtkio = "0.7.0-rc1"
 
