### 2025-08-14 cff8abc
add statement of need to README.MD
