### 2025-08-14 #8754 @hacknus
> I think I am done with my review for now.
>
> The project was really interesting to look at, especially to see how people are tackling Rust GUIs. My three unchecked points are all things that seem to be missing in the project itself.

Thank you very much @DanNixon! I updated the README with two introductory sentences about its statement of need and I added community guidelines. I also added some unit tests to the `math_tools.rs` and the different filters.
