### 2025-08-15 e9a4f4b
fix position display
