### 2025-08-15 #8754 @DanNixon
> > I think I am done with my review for now.
> > The project was really interesting to look at, especially to see how people are tackling Rust GUIs. My three unchecked points are all things that seem to be missing in the project itself.
>
> Thank you very much [@DanNixon](https://github.com/DanNixon)! I updated the README with two introductory sentences about its statement of need and I added community guidelines. I also added some unit tests to the `math_tools.rs` and the different filters.

Cool. I think everything looks good from my point of view then.
