### 2025-09-02 #8754 @hacknus
@fruzsinaagocs Do I have to put a declaration of the use of AI in the paper?
