### 2025-09-02 #8754 @fruzsinaagocs
@hacknus Thanks for the question! Although JOSS does not currently have a policy regarding the use of AI/LLMs yet, you as author are responsible for explaining the submitted code and its provenance. Adding a section on th use of AI would be helpful, especially for the reviewers' sake.
