### 2025-09-02 71abbae
add the declaration of the use of AI based tools
