### 2025-10-06 #8754 @jsdodge
I apologize again for the multiple delays and thank you for your patience. The software is an impressive contribution to a growing body of open-source software to support THz data analysis. I left a few items unchecked, for reasons that I explain below.

- I understand the JOSS criteria to require a statement of need in the documentation as well as in the paper. I would recommend including this at the very top of the README file, so that visitors can immediately see the purpose of the software. It can just restate what is in the paper.
- The documentation does not mention the `requirements.txt` file in the `scripts` folder, which includes the dependencies for the scripts. Moreover, the `thz-deconvolution` dependency has additional dependencies, `scikit-image` and `multiprocess`, which should be included in the `requirements.txt` file. Some instructions for installing these dependencies with `conda` and/or `pip` would be appropriate in this section. I was able to use `conda` successfully, with a `pip` installation of `thz-deconvolution` after adding its additional dependencies.
- Line 146 of `generate_psf.py` should point to `os.path.join(script_path, "../sample_data/psf.npz")`, not `os.path.join(script_path, "sample_data/psf.npz")`.
- The documentation needs some development to make it more accessible. It's unclear to me what the "Average in frequency domain" toggle button does, for example, and I was unable to discern how to change the axis limits for the plots. A good example of scientific GUI documentation is available in the [PIVA](https://piva.readthedocs.io/en/latest/) project, which provides a short description for each GUI element. A standalone documentation site with that level of detail is not necessary here, but a basic list in the README file that includes all of the GUI elements and their purposes would be helpful.
- There is no example to show a user how to use the *Optical Properties* panel. I tried selecting *Selected Pixel* from `resolutions_target_sample.thz` for the sample file and `reference.thz` for the reference, but nothing happened. Later, I discovered the following warning by accident in the *Settings* panel, which presumably means that the files did not have commensurate time samples.
    ```
    [WARN ] thz_image_explorer::data_thread: Time data from reference file does not match scan time data. Resizing and aligning reference signal. Phase data might not match correctly.
    [WARN ] thz_image_explorer::data_thread: Cannot align signals due to empty or invalid time data. Resizing naively.
    ```
- This is a minor issue that doesn't need to be addressed before publication, but in the *Pulse* tab, the labels for "DR" and "Water Lines" are overlapping and hard to read.
