### 2025-10-06 52b868f
 This is a combination of 2 commits.  This is the 1st commit message:
Track large thzimg files with Git LFS

implement feedback from reviewer 2

 The commit message #2 will be skipped:

 Track large thzimg files with Git LFS
