+++ README.MD
@@ -7,8 +7,10 @@
 
 <a href="https://github.com/unibe-icelab/thz-image-explorer/releases"><img src="assets/icons/icon.png" alt=“” width="100" height="100"> </img> </a>
 
-THz Image Explorer is an interactive cross-platform open-source tool for analyzing THz time-domain imaging data. Its goal is to
-make the analysis of THz scans easier and more accessible to researchers by providing a user-friendly interface and a set of powerful features.
+THz Image Explorer is an interactive cross-platform open-source tool for analyzing THz time-domain imaging data. Its
+goal is to
+make the analysis of THz scans easier and more accessible to researchers by providing a user-friendly interface and a
+set of powerful features.
 This Application is written in Rust and the GUI is based on egui and bevy.
 
 <p align="center">
@@ -78,7 +80,8 @@ sudo apt-get install libxcb-render0-dev libxcb-shape0-dev libxcb-xfixes0-dev lib
    ```sh
    sudo apt-get install libxcb-render0-dev libxcb-shape0-dev libxcb-xfixes0-dev libxkbcommon-dev libssl-dev pkg-config libx11-dev libasound2-dev libudev-dev libxkbcommon-x11-0
    ```
-   Note: For Fedora and other distributions consult [bevy](https://github.com/bevyengine) and [egui](https://github.com/emilk/egui) documentation for the required dependencies.
+   Note: For Fedora and other distributions consult [bevy](https://github.com/bevyengine)
+   and [egui](https://github.com/emilk/egui) documentation for the required dependencies.
 4. Install bundling tools:
     - Linux/macOS: `cargo install cargo-bundle`
     - Windows: `cargo install cargo-wix`
@@ -135,6 +138,28 @@ A reference file (standard `.thz`) can be loaded, which is used to compute the o
 first
 entry will be used.
 
+### Interface
+
+The interface is split into three main sections: Pulse View (main), Optical Properties View and 3D View.
+New files can be opened either via the `Open File` Menu in the left panel or by drag & drop of `.thz` or `.thzimg`
+files.
+A Meta Data editor is located on the left side and a filter configuration panel on the right side. These panels are
+visible in all three views.
+The main view shows the time domain trace and frequency domain spectrum of the selected pixel, as well as the 2D
+intensity plot of the scan.
+
+  <p align="center">
+    <img src="paper/thz_image_explorer_screenshot2.png" alt="Main View" width="800">
+  </p>
+
+The Optical Properties View shows the refractive index and absorption coefficient
+calculated from the selected pixel and the reference scan (this can either be different ROIs, the selected Pixel or the
+reference file).
+
+  <p align="center">
+    <img src="paper/thz_image_explorer_screenshot1.png" alt="Optical Properties View" width="800">
+  </p>
+
 ### dotTHz Meta Data Editor
 
 The meta-data editor allows the user to edit the meta-data of the loaded scan. The meta-data is stored in the `.thz` (or
@@ -168,6 +193,13 @@ displayed in the center plot.
 ### Filtering Pipeline
 
 The filtering process is a simple linear pipeline, where the output of one filter is the input of the next filter.
+On top of the filtering pipeline/panel general settings are located:
+
+- FFT Log Plot (default: on), which shows the frequency spectrum in logarithmic scale
+- Average in Frequency domain (default: off), which averages the selected pixels (ROI) in frequency domain instead of
+  time domain, this is useful if averaging in time domain leads to destructive interferences.
+- Downscale Factor (default: 1), which downscales the 2D image by averaging pixel blocks of size
+  `Downscale Factor x Downscale Factor`
 
 #### Time Domain Before FFT
 
@@ -177,7 +209,7 @@ Additionally, a simple band-pass filter can be applied to exclude secondary peak
 
 #### FFT
 
[…]
