### 2025-10-06 3185ca0
describe how to pan/zoom in plot
