+++ README.MD
@@ -146,7 +146,8 @@ files.
 A Meta Data editor is located on the left side and a filter configuration panel on the right side. These panels are
 visible in all three views.
 The main view shows the time domain trace and frequency domain spectrum of the selected pixel, as well as the 2D
-intensity plot of the scan.
+intensity plot of the scan. The plots can be zoomed by right-clicking and dragging a rectangle, panning can be done by
+left-clicking and dragging the mouse. Double-clicking will reset the zoom and pan.
 
   <p align="center">
     <img src="paper/thz_image_explorer_screenshot2.png" alt="Main View" width="800">
