### 2025-10-06 34d989b
default avg in time domain
