### 2025-10-07 808f5d3
add more info about navigating in plots
