+++ scripts/requirements.txt
@@ -1,5 +1,5 @@
-numpy==1.26.4
-tqdm==4.66.4
-thz-deconvolution==1.0.4
-matplotlib~=3.10.3
-shapely==2.0.1
+numpy
+tqdm
+thz-deconvolution>=1.0.4
+matplotlib
+shapely
