### 2025-10-07 #8754 @hacknus
Thanks @jsdodge for opening issues for these two problems. I implemented a fix for the `requirements.txt`, a fix for the zoom would have to be implement in one of the dependencies (egui-plot), so I agree, this goes on the "future work" pile.
