### 2025-10-15 #8754 @fruzsinaagocs
@jsdodge Thank you for your review! It looks like the remaining unchecked items on your list relate to functionality and functionality documentation. What's missing here?
I also see two remaining open issues, one in connection with dependencies and one about zooming in/out. Have the dependency issues been addressed?
