+++ Cargo.toml
@@ -82,8 +82,6 @@ egui-file-dialog = { git = "https://github.com/hacknus/egui-file-dialog", branch
 [target.'cfg(target_os = "macos")'.dependencies]
 rfd = "0.15"
 futures = "0.3.31"
-
-[target.'cfg(target_os = "macos")'.dependencies]
 tracing = "0.1.41"
 objc2 = "0.6.1"
 objc2-app-kit = { version = "0.3.1" }
