+++ scripts/generate_psf.py
@@ -1,171 +1,171 @@
-from thz_deconvolution import *
 import argparse
-import numpy as np
 import os
+
 import matplotlib.pyplot as plt
+import numpy as np
+from thz_deconvolution import *
 
-np.set_printoptions(precision=4)
+if __name__ == "__main__":
+    np.set_printoptions(precision=4)
 
-show = False
+    show = False
 
-# Frequency and filter parameters
-low_cut = 0.15
-high_cut = 10.0
-start_freq = 0.2
-end_freq = 4.0
-win_width = 0.5
-n_filters = 20
-w_max = 30
+    # Frequency and filter parameters
+    low_cut = 0.15
+    high_cut = 10.0
+    start_freq = 0.2
+    end_freq = 4.0
+    win_width = 0.5
+    n_filters = 20
+    w_max = 30
 
-print("* Fit parameters: ")
-print("  - " + str(start_freq) + " THz to " + str(end_freq) + " THz")
-print("  - " + str(n_filters) + " filters")
-print("  - " + str(w_max) + " mm max beam width")
-print("  - " + str(win_width) + " mm window width")
-print()
-
-script_path = os.path.dirname(os.path.abspath(__file__))
-
-# Argument parsing
-parser = argparse.ArgumentParser(description="Generate PSF using raw measurements.")
-parser.add_argument(
-    "--path_x",
-    required=True,
-    help="Path to the knife edge measurement file in x."
-)
-parser.add_argument(
-    "--path_y",
-    required=True,
-    help="Path to the knife edge measurement file in y."
-)
-args = parser.parse_args()
-
-x_path = args.path_x
-y_path = args.path_y
-
-# Load the raw PSF measurements
-print()
-print("* Loading knife edge measurements")
-
-x_psf_lr, y_psf_lr, np_psf_t_x_lr, np_psf_t_y_lr, times_psf = load_knife_edge_meas(x_path, y_path)
-
-
-x_psf_lr = np.split(x_psf_lr, 2)
-y_psf_lr = np.split(y_psf_lr, 2)
-np_psf_t_x_lr = np.split(np_psf_t_x_lr, 2)
-np_psf_t_y_lr = np.split(np_psf_t_y_lr, 2)
-
-# Flipping the left part
-x_psf_lr[0] = -np.flip(x_psf_lr[0])
-y_psf_lr[0] = -np.flip(y_psf_lr[0])
-np_psf_t_x_lr[0] = np.flip(np_psf_t_x_lr[0])
-np_psf_t_y_lr[0] = np.flip(np_psf_t_y_lr[0])
-
-popt_xs_lr = []
-popt_ys_lr = []
-
-for x_psf, y_psf, np_psf_t_x, np_psf_t_y in zip(x_psf_lr, y_psf_lr, np_psf_t_x_lr, np_psf_t_y_lr):
-
-    x_psf -= np.mean(x_psf)
-    y_psf -= np.mean(y_psf)
-
+    print("* Fit parameters: ")
+    print("  - " + str(start_freq) + " THz to " + str(end_freq) + " THz")
+    print("  - " + str(n_filters) + " filters")
+    print("  - " + str(w_max) + " mm max beam width")
+    print("  - " + str(win_width) + " mm window width")
     print()
-    print("* Fitting the mean PSF")
-    n_min = 0
-    n_max = -1
-    x0, y0, popt_x, popt_y = fit_mean_beam(
-        x_psf, y_psf, np_psf_t_x, np_psf_t_y, [n_min, n_max], plot=show)
 
-    # Create the PSF
-    x_start = np.abs(x_psf[0])
-    y_start = np.abs(y_psf[0])
[…]
