### 2025-10-16 eab46b0
reduce npz size
