+++ .github/workflows/deployment.yml
@@ -26,37 +26,48 @@ jobs:
             target: x86_64-pc-windows-msvc
           - os: windows-2022
             target: aarch64-pc-windows-msvc
+
     runs-on: ${{ matrix.os }}
     steps:
       - uses: actions/checkout@v4
 
       - name: Install Dependencies (Linux)
         if: contains(matrix.os, 'ubuntu')
-        run: sudo apt-get update && sudo apt-get install -y libclang-dev libgtk-3-dev libxcb-render0-dev libxcb-shape0-dev libxcb-xfixes0-dev libxkbcommon-dev libssl-dev libudev-dev && sudo apt-get install -y libasound2-dev && sudo apt-get install -y cmake && cargo install --git https://github.com/burtonageo/cargo-bundle
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y libclang-dev libgtk-3-dev libxcb-render0-dev \
+            libxcb-shape0-dev libxcb-xfixes0-dev libxkbcommon-dev libssl-dev \
+            libudev-dev libasound2-dev cmake
+          cargo install --git https://github.com/burtonageo/cargo-bundle
 
       - name: Install Dependencies (macOS)
         if: contains(matrix.os, 'macos')
-        run: brew install cmake && cargo install --git https://github.com/burtonageo/cargo-bundle
+        run: |
+          brew install cmake
+          cargo install --git https://github.com/burtonageo/cargo-bundle
 
       - name: Install Dependencies (Windows)
         if: contains(matrix.os, 'windows')
-        run: cargo install --force cargo-wix && choco install cmake --installargs 'ADD_CMAKE_TO_PATH=System' -y
-
-      - name: Set CARGO_FEATURES environment variable (Windows)
-        if: contains(matrix.os, 'windows')
         run: |
-          echo "CARGO_FEATURES=self_update" >> $env:GITHUB_ENV
+          cargo install --force cargo-wix
+          choco install cmake --installargs 'ADD_CMAKE_TO_PATH=System' -y
+
+      - name: Set CARGO_FEATURES (Windows)
+        if: contains(matrix.os, 'windows')
+        run: echo "CARGO_FEATURES=self_update" >> $env:GITHUB_ENV
 
       - name: Build
-        run: cargo build --features self_update --release
+        run: |
+          rustup target add ${{ matrix.target }}
+          cargo build --release --features self_update --target ${{ matrix.target }}
 
       - name: Build .deb Package (Linux)
         if: contains(matrix.os, 'ubuntu')
-        run: cargo bundle --features self_update --release
+        run: cargo bundle --features self_update --release --target ${{ matrix.target }}
 
       - name: Build .app Package (macOS)
         if: contains(matrix.os, 'macos')
-        run: cargo bundle --features self_update --release
+        run: cargo bundle --features self_update --release --target ${{ matrix.target }}
 
       - name: Build .msi Package (Windows)
         if: contains(matrix.os, 'windows')
@@ -66,7 +77,9 @@ jobs:
       - name: Compress Output (Linux Binary)
         if: contains(matrix.os, 'ubuntu')
         run: |
-          cd target/release
+          BIN_DIR="target/${{ matrix.target }}/release"
+          [ -d "$BIN_DIR" ] || BIN_DIR="target/release"
+          cd "$BIN_DIR"
           zip -r thz-image-explorer-${{ matrix.target }}.zip thz-image-explorer
           mv thz-image-explorer-${{ matrix.target }}.zip $GITHUB_WORKSPACE/
 
@@ -74,7 +87,9 @@ jobs:
[…]
