### 2025-11-04 fe45e6c
implement inputs from editor
