+++ paper/paper.bib
@@ -6,6 +6,7 @@
     number={3},
     pages={103--104},
     year={2014},
+	doi = {https://doi.org/10.1145/2692956.2663188},
     organization={ACM}
 }
 
@@ -81,16 +82,6 @@
 	year = {2023},
 	pages = {795--813},
 	file = {PDF:/Users/linus/Zotero/storage/IMCVJTKU/Lee et al. - 2023 - The dotTHz Project A Standard Data Format for Terahertz Time-Domain Data.pdf:application/pdf},
-}
-
-@book{shreiner_opengl_2009,
-	edition = {7th},
-	title = {{OpenGL} {Programming} {Guide}: {The} {Official} {Guide} to {Learning} {OpenGL}, {Versions} 3.0 and 3.1},
-	isbn = {0-321-55262-8},
-	abstract = {OpenGL is a powerful software interface used to produce high-quality, computer-generated images and interactive applications using 2D and 3D objects, bitmaps, and color images. The OpenGLProgramming Guide, Seventh Edition, provides definitive and comprehensive information on OpenGL and the OpenGL Utility Library. The previous edition covered OpenGL through Version 2.1. This seventh edition of the best-selling red book describes the latest features of OpenGL Versions 3.0 and 3.1. You will find clear explanations of OpenGL functionality and many basic computer graphics techniques, such as building and rendering 3D models; interactively viewing objects from different perspective points; and using shading, lighting, and texturing effects for greater realism. In addition, this book provides in-depth coverage of advanced techniques, including texture mapping, antialiasing, fog and atmospheric effects, NURBS, image processing, and more. The text also explores other key topics such as enhancing performance, OpenGL extensions, and cross-platform techniques. This seventh edition has been updated to include the newest features of OpenGL Versions 3.0 and 3.1, including Using framebuffer objects for off-screen rendering and texture updates Examples of the various new buffer object types, including uniform-buffer objects, transform feedback buffers, and vertex array objects Using texture arrays to increase performance when using numerous textures Efficient rendering using primitive restart and conditional rendering Discussion of OpenGLs deprecation mechanism and how to verify your programs for future versions of OpenGL This edition continues the discussion of the OpenGL Shading Language (GLSL) and explains the mechanics of using this language to create complex graphics effects and boost the computational power of OpenGL. The OpenGL Technical Library provides tutorial and reference books for OpenGL. The Library enables programmers to gain a practical understanding of OpenGL and shows them how to unlock its full potential. Originally developed by SGI, the Library continues to evolve under the auspices of the Khronos OpenGL ARB Working Group, an industry consortium responsible for guiding the evolution of OpenGL and related technologies.},
-	publisher = {Addison-Wesley Professional},
-	author = {Shreiner, Dave and Group, The Khronos OpenGL ARB Working},
-	year = {2009},
 }
 
 @article{koch-dandolo_reflection_2015,
