+++ paper/paper.md
@@ -31,17 +31,16 @@ bibliography: paper.bib
 
 THz time-domain spectroscopy (TDS) is a fast-growing field with applications to perform non-destructive studies of
 material properties [@neu_tutorial_2018].
-The pulses can either be measured after passing through (transmission spectrum) or after being reflected (reflection
-spectrum) by a sample and are recorded in the time domain. Through Fourier analysis (FFT), we can investigate the
+The pulses can either be measured after passing through (transmission spectrum) or after being reflected by (reflection
+spectrum) a sample. Through Fourier analysis (FFT), we can investigate the
 complex refractive index and absorption coefficient of the sample.
 By placing either the sample or the optical setup on a moving stage the sample can be imaged in 2D.
 
 ![THz Image Explorer icon.\label{fig:icon}](icon.png){#id .class width=20%}
 
 We developed an interactive graphical user interface (GUI), written
-in [Rust](https://www.rust-lang.org) [@matsakis2014rust], to aid
-investigating acquired 2D scans. The
-application implements the dotTHz standard [@lee_dotthz_2023] and is platform independent and open-source, thus making
+in [Rust](https://www.rust-lang.org) [@matsakis2014rust], to aid in investigating acquired 2D scans. The
+application implements the dotTHz standard [@lee_dotthz_2023] and is platform independent and open-source, making
 it easier to maintain and increasing its reach.
 
 ![THz Image Explorer screenshot.\label{fig:screenshot}](screenshot.png)
@@ -54,23 +53,23 @@ Commercial suppliers provide closed-source analysis tools (e.g. [Menlo Systems](
 code cannot be adapted by the user, which is often essential in research environments and extends the maintainability of
 the code.
 Solutions published by the scientific community are not available on all platforms, are only applicable on single pixel
-measurements and/or not focused on an interactive workflow [@peretti_thz-tds_2019; @loaiza_thztools_2024].
-With this application, we provide a high-performance solution written in Rust, that allows an interactive analysis of 2D
+measurements and/or are not focused on an interactive workflow [@peretti_thz-tds_2019; @loaiza_thztools_2024].
+With this application, we provide a high-performance solution written in Rust that allows an interactive analysis of 2D
 THz
 scans with multiple filters and a 3D viewer.
-This work is open-source and pre-built bundles are available for Linux, macOS and Windows, thus available to the
-entire scientific community.
+The work is open-source, and pre-built bundles are provided for Linux, macOS, and Windows, ensuring broad accessibility
+for the scientific community.
 
 # Structure
 
 The application is multithreaded with two main threads:
 
-- GUI thread
-- Data thread
+- GUI thread,
+- Data thread.
 
 The GUI uses [egui](https://www.egui.rs), an immediate-mode GUI library for rust.
 
-The structure of the software architecture is shown in figure \ref{fig:software-architecture}.
+The structure of the software is shown in figure \ref{fig:software-architecture}.
 
 ![Software Architecture.\label{fig:software-architecture}](thz-image-explorer.drawio.png){width=80% .center}
 
@@ -81,19 +80,17 @@ using the COCoNuT setup [@coconut_2025].
 
 ## Optical Properties Calculation
 
-The refractive index $n$ and absorption coefficient $\alpha$ can be computed using the relations from [@Jepsen2019]. The
-user can select a source and a reference for the calculation.
+The user can select a source and reference scan in the drop-down menu, after which the refractive index $n$ and
[…]
