### 2025-11-04 #8754 @hacknus
> Please also make the following corrections to the bibliography:
>
> * Please convert the bevy engine reference to a hyperlink to their github, as it is not clear how best to cite this resource.

Thanks a lot for the inputs! I have implemented all of the above. Regarding the bevy citation, I have based the citation on this discussion here: https://github.com/bevyengine/bevy/discussions/8098

Let me know if I should still change it to a hyperlink instead of a citation.
