### 2025-11-05 797236a
minor fix for GUI alignment
