### 2025-11-05 #8754 @hacknus
Thanks @fruzsinaagocs !

tag: [1.0.1](https://github.com/unibe-icelab/thz-image-explorer/tree/1.0.1)
DOI: https://doi.org/10.5281/zenodo.17529048
