+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "thz-image-explorer"
 authors = ["Linus Leo Stöckli"]
-version = "1.0.0"
+version = "1.0.1"
 edition = "2021"
 description = "Explorer Software for THz TDS images."
 license = "GPL-3.0"
