+++ paper/paper.bib
@@ -6,7 +6,7 @@
     number={3},
     pages={103--104},
     year={2014},
-	doi = {https://doi.org/10.1145/2692956.2663188},
+	doi = {10.1145/2692956.2663188},
     organization={ACM}
 }
 
