### 2025-11-07 5a27605
changes in paper.md
