### 2025-11-07 2dbd09a
changes in paper.bib
