+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @inproceedings{matsakis2014rust,
-    title={The rust language},
+    title={The {R}ust language},
     author={Matsakis, Nicholas D and Klock II, Felix S},
     booktitle={ACM SIGAda Ada Letters},
     volume={34},
@@ -12,7 +12,7 @@
 
 @software{bevyengine,
     title = {Bevy Engine},
-    author = {Bevy Contributors},
+    author = {{Bevy Contributors}},
     version = {0.16.1},
     date = {2025-05-31},
     abstract = {A refreshingly simple data-driven game engine built in Rust},
@@ -66,7 +66,7 @@
 
 
 @article{lee_dotthz_2023,
-	title = {The {dotTHz} {Project}: {A} {Standard} {Data} {Format} for {Terahertz} {Time}-{Domain} {Data}},
+	title = {The {dotTHz} Project: A Standard Data Format for Terahertz Time-Domain Data},
 	volume = {44},
 	issn = {1866-6892, 1866-6906},
 	shorttitle = {The {dotTHz} {Project}},
@@ -100,7 +100,7 @@
 }
 
 @article{loaiza_thztools_2024,
-	title = {{THzTools}: data analysis software for terahertztime-domain spectroscopy},
+	title = {{THzTools}: data analysis software for terahertz time-domain spectroscopy},
 	volume = {9},
 	copyright = {http://creativecommons.org/licenses/by/4.0/},
 	issn = {2475-9066},
@@ -139,7 +139,7 @@
 }
 
 @article{peretti_thz-tds_2019,
-	title = {{THz}-{TDS} {Time}-{Trace} {Analysis} for the {Extraction} of {Material} and {Metamaterial} {Parameters}},
+	title = {{THz}-{TDS} Time-Trace Analysis for the Extraction of Material and Metamaterial Parameters},
 	volume = {9},
 	copyright = {https://ieeexplore.ieee.org/Xplorehelp/downloads/license-information/IEEE.html},
 	issn = {2156-342X, 2156-3446},
