### 2025-11-07 07db237
fixing bad phrasing in paper.md
