### 2025-11-07 #8754 @danielskatz
👋 @hacknus - As track editor, I found a number of things in the paper that need to be changed for JOSS publication, as shown in https://github.com/unibe-icelab/thz-image-explorer/pull/16.  Please merge this, or let me know what you disagree with, then we can proceed to acceptance and publication.
