### 2025-11-07 #8754 @hacknus
Hi @danielskatz, thanks a lot for reading through the manuscript! I left two comments about the citation style, if is fine with you to drop these changes then I can merge the PR.
