### 2025-11-07 #8754 @danielskatz
I've responded that those two changes are correct because those references are being mentioned directly, not just indirectly to support other text.
