### 2025-11-07 999e504
add citation file
